Move psql script generation code to submodule.
