+++ src/tasks/dump_db.rs
@@ -1,5 +1,4 @@
 use std::{
-    collections::{BTreeMap, VecDeque},
     fs::File,
     path::{Path, PathBuf},
 };
@@ -23,10 +22,9 @@ pub fn dump_db(
     defer! {{
         std::fs::remove_dir_all(&export_dir).unwrap();
     }}
-    let visibility_config: VisibilityConfig = toml::from_str(include_str!("dump-db.toml")).unwrap();
     let export_script = export_dir.join("export.sql");
     let import_script = export_dir.join("import.sql");
-    visibility_config.gen_psql_scripts(&export_script, &import_script)?;
+    gen_scripts::gen_scripts(&export_script, &import_script)?;
     run_psql(&database_url, &export_script)?;
     let tarball = create_tarball(&export_dir)?;
     defer! {{
@@ -35,138 +33,6 @@ pub fn dump_db(
     upload_tarball(&tarball, &target_name, &env.uploader)?;
     println!("Database dump uploaded to {}.", &target_name);
     Ok(())
-}
-
-/// An enum indicating whether a column is included in the database dumps.
-/// Public columns are included, private are not.
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
-enum ColumnVisibility {
-    Private,
-    Public,
-}
-
-/// Filtering information for a single table. The `dependencies` field is only
-/// used to determine the order of the tables in the generated import script,
-/// and should list all tables the current tables refers to with foreign key
-/// constraints on public columns. The `filter` field is a valid SQL expression
-/// used in a `WHERE` clause to filter the rows of the table. The `columns`
-/// field maps column names to their respective visibilities.
-#[derive(Clone, Debug, Deserialize)]
-struct TableConfig {
-    #[serde(default)]
-    dependencies: Vec<String>,
-    filter: Option<String>,
-    columns: BTreeMap<String, ColumnVisibility>,
-}
-
-/// Subset of the configuration data to be passed on to the Handlbars template.
-#[derive(Debug, Serialize)]
-struct HandlebarsTableContext<'a> {
-    name: &'a str,
-    filter: Option<&'a str>,
-    columns: String,
-}
-
-impl TableConfig {
-    fn handlebars_context<'a>(&'a self, name: &'a str) -> Option<HandlebarsTableContext<'a>> {
-        let columns = self
-            .columns
-            .iter()
-            .filter(|&(_, &vis)| vis == ColumnVisibility::Public)
-            .map(|(col, _)| format!("\"{}\"", col))
-            .collect::<Vec<String>>()
-            .join(", ");
-        if columns.is_empty() {
-            None
-        } else {
-            Some(HandlebarsTableContext {
-                name,
-                filter: self.filter.as_ref().map(String::as_str),
-                columns,
-            })
-        }
-    }
-}
-
-/// Maps table names to the respective configurations. Used to load `dump_db.toml`.
-#[derive(Clone, Debug, Deserialize)]
-#[serde(transparent)]
-struct VisibilityConfig(BTreeMap<String, TableConfig>);
-
-/// Subset of the configuration data to be passed on to the Handlbars template.
-#[derive(Debug, Serialize)]
-struct HandlebarsContext<'a> {
-    tables: Vec<HandlebarsTableContext<'a>>,
-}
-
-impl VisibilityConfig {
-    /// Sort the tables in a way that dependencies come before dependent tables.
-    ///
-    /// Returns a vector of table names.
-    fn topological_sort(&self) -> Vec<&str> {
-        let mut result = Vec::new();
-        let mut num_deps = BTreeMap::new();
-        let mut rev_deps: BTreeMap<_, Vec<_>> = BTreeMap::new();
-        for (table, config) in self.0.iter() {
-            num_deps.insert(table.as_str(), config.dependencies.len());
-            for dep in &config.dependencies {
-                rev_deps.entry(dep.as_str()).or_default().push(table.as_str());
-            }
-        }
-        let mut ready: VecDeque<&str> = num_deps
-            .iter()
-            .filter(|(_, &count)| count == 0)
-            .map(|(&table, _)| table)
-            .collect();
-        while let Some(table) = ready.pop_front() {
-            result.push(table);
-            for dep in rev_deps.get(table).iter().copied().flatten() {
-                *num_deps.get_mut(dep).unwrap() -= 1;
-                if num_deps[dep] == 0 {
-                    ready.push_back(dep);
-                }
-            }
-        }
-        assert_eq!(
-            self.0.len(),
-            result.len(),
-            "circular dependencies in database dump configuration detected",
-        );
-        result
-    }
-
-    fn handlebars_context(&self) -> HandlebarsContext<'_> {
-        let tables = self
-            .topological_sort()
-            .into_iter()
-            .filter_map(|table| self.0[table].handlebars_context(table))
-            .collect();
-        HandlebarsContext { tables }
-    }
-
-    fn gen_psql_scripts(
-        &self,
-        export_script: &Path,
-        import_script: &Path,
-    ) -> Result<(), PerformError> {
-        let context = self.handlebars_context();
-        let mut handlebars = handlebars::Handlebars::new();
-        handlebars.register_escape_fn(handlebars::no_escape);
-        let export_sql = File::create(export_script)?;
-        handlebars.render_template_to_write(
-            include_str!("dump-export.hbs"),
-            &context,
-            export_sql,
-        )?;
-        let import_sql = File::create(import_script)?;
-        handlebars.render_template_to_write(
-            include_str!("dump-import.hbs"),
-            &context,
-            import_sql,
-        )?;
-        Ok(())
-    }
 }
 
 fn run_psql(database_url: &str, export_script: &Path) -> Result<(), PerformError> {
@@ -223,77 +89,4 @@ fn upload_tarball(
     Ok(())
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_util::pg_connection;
-    use diesel::prelude::*;
-    use std::collections::HashSet;
-
-    /// Test whether the schema in the visibility configuration matches the test database.
-    #[test]
-    fn check_visibility_config() {
-        let conn = pg_connection();
-        let db_columns: HashSet<_> = get_db_columns(&conn)
-            .into_iter()
-            .map(|c| (c.table_name, c.column_name))
-            .collect();
-        let visibility_config: VisibilityConfig =
-            toml::from_str(include_str!("dump-db.toml")).unwrap();
-        let vis_columns: HashSet<_> = visibility_config
-            .0
-            .iter()
-            .flat_map(|(table, config)| {
-                config
-                    .columns
-                    .iter()
-                    .map(move |(column, _)| (table.clone(), column.clone()))
-            })
-            .collect();
-        let mut errors = vec![];
-        for (table, col) in db_columns.difference(&vis_columns) {
-            errors.push(format!(
-                "No visibility information for columns {}.{}.",
-                table, col
-            ));
-        }
-        for (table, col) in vis_columns.difference(&db_columns) {
-            errors.push(format!(
-                "Column {}.{} does not exist in the database.",
-                table, col
-            ));
-        }
-        assert!(
-            errors.is_empty(),
-            "The visibility configuration does not match the database schema:\n{}",
-            errors.join("\n  - "),
-        );
-    }
-
-    mod information_schema {
-        table! {
-            information_schema.columns (table_schema, table_name, column_name) {
-                table_schema -> Text,
-                table_name -> Text,
-                column_name -> Text,
-                ordinal_position -> Integer,
-            }
-        }
-    }
-
-    #[derive(Debug, PartialEq, Queryable)]
-    struct Column {
-        table_name: String,
-        column_name: String,
-    }
-
-    fn get_db_columns(conn: &PgConnection) -> Vec<Column> {
-        use information_schema::columns::dsl::*;
-        columns
-            .select((table_name, column_name))
-            .filter(table_schema.eq("public"))
-            .order_by((table_name, ordinal_position))
-            .load(conn)
-            .unwrap()
-    }
-}
+mod gen_scripts;
